feature-processor/requirements.txt
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-json-logger==2.0.7
flask==3.0.0
pyyaml==6.0.1
numpy==1.24.4
scipy==1.11.4
